==> 2_Ratings_Matrix.ipynb <==
# Building a Ratings Matrix

The purpose of this workbook is to build a user-item ratings matrix which will be used for a collaborative based recommender system.

The final output will be a ratings matrix which will be saved as a .pkl (pickle) file to be used in future workbooks.

***

# Import Python libraries as needed
import numpy as np
import pandas as pd

# Read in saved data-frames from previous workbook
business = pd.read_pickle('data/user/business.pkl')
review = pd.read_pickle('data/user/review.pkl')

# Quick overview of business table
print(business.shape)
business.head(3)

# Quick overview of review table
print(review.shape)
review.head(3)

# Confirm there are no records in the review table with missing star ratings
review['stars'].isna().value_counts()

***

# Create a 0-n index that can be called as a series for each unique 'user_id' value
unique_user = review['user_id'].unique()

# Transform series into a data-frame
unique_user = pd.DataFrame(unique_user)

# Reset the index such that the index is copied into a different column labelled 'user_num'
unique_user = unique_user.reset_index().rename(columns = {'index':'user_num',0:'user_id'})

# Review final result
unique_user

# Save unique_user df to pickle file for use in future workbook
unique_user.to_pickle('data/user/unique_user.pkl')

# Create a 0-n index that can be called as a series for each unique 'business_id' value
unique_business = review['business_id'].unique()

# Transform series into a data-frame
unique_business = pd.DataFrame(unique_business)

# Reset the index such that the index is copied into a different column labelled 'business_num'
unique_business = unique_business.reset_index().rename(columns = {'index':'business_num',0:'business_id'})

# Review final result
unique_business

# Save unique_user df to pickle file for use in future workbook
unique_business.to_pickle('data/user/unique_business.pkl')

# Capture the length of each unique user and business lists
num_user = len(unique_user)
num_business = len(unique_business)

# Print both values
print('unique users:', num_user, '\nunique businesses:', num_business)

%%time
# Capture cell execution time (approx. 1 min.)

# Create an empty array with a size that is equal to number of unique users x number of unique businesses
ratings_array = np.full((num_user, num_business), np.nan)

# Review final result
ratings_array

# Create seperate data-frame for user-item ratings matrix
mini_review = review.loc[:,['user_id','business_id','stars','date']].copy()

# Sort data frame by most recent to least recent date (in order to assist with dropping duplicates)
mini_review.sort_values(by='date', ascending = False, inplace = True)

# Quick overview of the 'mini_review' data-frame
print(mini_review.shape)
mini_review.head(3)

# Drop all duplicate values such that there is only one star rating for every unique user-business combination
# By default the first value is kept for each set of duplicate values
# Since we previously sorted the review by date/time, we can be assured that only the most recent review for each business has been kept for each user
mini_review.drop_duplicates(['user_id','business_id'], inplace = True)
mini_review.shape

# Lookup the unique integer based index values for each unique user_id and business_id
mini_review = pd.merge(mini_review, unique_user, how = 'left')
mini_review = pd.merge(mini_review, unique_business, how = 'left')
mini_review.head(3)

# Confirm that user_num and business_num values have been applied successfully and that there are no missing values
print(mini_review['business_num'].isna().value_counts(),'\n')
print(mini_review['user_num'].isna().value_counts())

# Showing an example record by one user designated as 'user_num 28532'
mini_review[mini_review['user_num'] == 28532]

# Transfer the star rating from the 'mini_review' df into the 'ratings_array' df

# Loop over the number of rows within the 'mini_review' data-frame
for row in mini_review.itertuples():
    
    # Capture the value under the 'user_num' column
    user_index = row[5]
    
    # Capture the value under the 'business_num' column
    business_index = row[6]
    
    # Capture the rating under the 'stars' column
    rating_value = row[3]
    
    # Update the 'ratings_array' using the 'user_num' and 'business_num' as the index value in order to store the 'stars' value
    ratings_array[user_index, business_index] = rating_value
    
# Transform the 'ratings_array' into a data-frame and then output a few of its records    
ratings_matrix = pd.DataFrame(ratings_array)
ratings_matrix

# Select any 'user_num' value to act as the selected user (for use in future functions)
current_user = 5

# Create a boolean series indicating the non-null columns for the given row number (i.e. the 'current_user')
user_pick_non_blanks = ratings_df.loc[current_user,].isna() == False

# Pull all the non-null values from the ratings table for the 'current_user'
ratings_matrix.loc[current_user,][user_pick_non_blanks]

# Save the 'ratings_matrix' data-frame to a .pkl file
ratings_matrix.to_pickle('data/user/ratings_matrix.pkl')
